fix(zsh): call register-python-argcomplete only on non-homebrew systems
There is a weird difference between pipx with homebrew and pipx on
archlinux. Apparently archlinux uses python-argcomplete as dependency
and hence the autocompletion is handled through that. Therefore, on that
system you need to call the function register-python-argcomplete. On
homebrew on the other hand, python-argcomplete is NOT a dependency and
completion is installed automatically as a static function. Hence, on
mac you should not and cannot call the register-python-argcomplete
function.